+++ src/components/app-containers/SrPage/SrPage.vue
@@ -2,14 +2,6 @@
   <div v-bind:[attribute]="value">
     <div class="new-style">
       <v-app class="sr-page">
-        <div>
-          <div class="breadcrumb-container">
-            <sr-breadcrumbs />
-          </div>
-          <div class="help-btn-container">
-            <sr-help-button />
-          </div>
-        </div>
         <v-main>
           <slot />
         </v-main>